统一修改搜索为变量名为searchText

+++ tmgg-system/src/main/java/io/tmgg/modules/asset/service/SysAssetService.java
@@ -29,7 +29,9 @@ public class SysAssetService extends BaseService<SysAsset> {
     public void deleteById(String id) {
         SysAsset a = sysAssetDao.findOne(id);
         if(a.getType() ==1){
-            sysFileService.deleteById(a.getContent());
+            if(a.getContent() != null){
+                sysFileService.deleteById(a.getContent());
+            }
         }
         super.deleteById(id);
     }
